EID-476 - Create Stub country cert with shorter expiry date
-Currently all certs are created using the same expiry date. Where
there is a certificate chain, we need to be able to test that the cert
with the shortest expiry date is compared to the trust anchor
refresh delay and not the first cert in the chain.
-This cert was generated by temporarily changing the cfssl-config.json file
to use a shorter expiry date. This cert is not permanently added to
generate-certs.sh so the expiry date always remains shorter than the others.

Solo - @JHjava